# Conflicts:
#	app/gui2/package-lock.json
#	app/gui2/package.json
#	app/gui2/src/components/GraphEditor.vue
#	app/gui2/src/main.ts

+++ engine/runtime/src/test/java/org/enso/interpreter/test/MetaObjectTest.java
@@ -188,7 +188,7 @@ public class MetaObjectTest extends TestBase {
       }
       switch (t.getMetaSimpleName()) {
         // represented as primitive values without meta object
-        case "Decimal" -> {}
+        case "Float" -> {}
         // has no instances
         case "Array_Proxy" -> {}
         // Warning is transparent and invisible
